Simplify collectInto generic type parameter
Motivation
----------

The collectInto methods allow callers the ability to pass their own type to collect their results into. The type will use a CollectorFactory to lookup a Collector for the appropriate return type. Currently, the collectInto methods require the given type to be a subtype of the JDK Collection type. Not all collections extend from this interface (e.g. Eclipse Collections) so having the generic type parameter require this makes the method call more restrictive than it has to be. Relaxing the parameter type to just the return type allows for Collectors that don't leverage the JDK collection to be used.

Modifications
-------------

The generic type parameter has been simplified from <R extends Collection<? super T>> to <R>. This is also in line with how the ResultBearing interface handles it's collectInto methods.

Result
------

The result is the collectInto methods on ResultIterable can support collectors that don't use the JDK Collection type.